Tratamento de excecoes - Not Found e Referential Integrity

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFindException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -16,23 +21,23 @@ public class ProductService {
     private ProductRepository repository;
 
     @Transactional(readOnly = true)
-    public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+    public ProductDTO findById(Long id) {
+        Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFindException("Resource not found!"));
         return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
-    public Page<ProductDTO> findAll(Pageable pageable){
+    public Page<ProductDTO> findAll(Pageable pageable) {
         Page<Product> result = repository.findAll(pageable);
         return result.map(x -> new ProductDTO(x));
 
     }
 
     @Transactional
-    public ProductDTO insert(ProductDTO dto){
+    public ProductDTO insert(ProductDTO dto) {
 
         Product entity = new Product();
-        CopyDtoToEntity(dto,entity);
+        CopyDtoToEntity(dto, entity);
 
         entity = repository.save(entity);
         return new ProductDTO(entity);
@@ -40,20 +45,34 @@ public class ProductService {
     }
 
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto){
+    public ProductDTO update(Long id, ProductDTO dto) {
+        try {
+            Product entity = repository.getReferenceById(id);
+            CopyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+             throw new ResourceNotFindException("Resource not find!");
 
-        Product entity = repository.getReferenceById(id);
-        CopyDtoToEntity(dto,entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        }
+
     }
 
-    @Transactional
-    public void delete(Long id){
-         repository.deleteById(id);;
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)){
+            throw new ResourceNotFindException("Resource not find!");
+        }
+        try {
+            repository.deleteById(id);
+        } catch(DataIntegrityViolationException e){
+            throw new DatabaseException("Referential Integrity failure!");
+
+        }
+        ;
     }
 
-    private void CopyDtoToEntity(ProductDTO dto, Product entity){
+    private void CopyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
         entity.setDescription(dto.getDescription());
         entity.setPrice(dto.getPrice());
